chore: Update displayDate type to string in GameData; refactor the fetching on main page
