+++ src/types.ts
@@ -5,5 +5,5 @@ export type GameData = {
   answers: string[];
   count: number;
   gameNumber: number;
-  displayDate: Date;
+  displayDate: string;
 };
